% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-launcher.R
\name{kgschart_app}
\alias{kgschart_app}
\title{Launch shiny app}
\usage{
kgschart_app(...)
}
\arguments{
\item{...}{optional arguments for \code{\link[shiny]{runApp}}}
}
\description{
Start "KGS Rank Graph Parser" applicaiton.
}
\examples{
\dontrun{
kgschart_app()}
}

