% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{plot.kgschart}
\alias{plot.kgschart}
\title{Plot Method for kgschart object}
\usage{
\method{plot}{kgschart}(x, y = NULL, image = FALSE, ...)
}
\arguments{
\item{x}{\code{kgschart} class object}

\item{y}{not in use}

\item{image}{if TRUE, plot the image, otherwise plot the data}

\item{...}{not in use}
}
\value{
\code{gtable} object if \code{image} is TRUE,
\code{ggplot} object otherwise
}
\description{
Plot Method for kgschart object
}

