% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{print.kgschart}
\alias{print.kgschart}
\title{Print method for kgschart object}
\usage{
\method{print}{kgschart}(x, ...)
}
\arguments{
\item{x}{\code{kgschart} class object}

\item{...}{not in use}
}
\value{
Nothing
}
\description{
Display basic information
}

