% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{kgschart}
\alias{kgschart}
\title{KGS Rank Graph Parser}
\usage{
kgschart(src, keep_image = TRUE, ...)
}
\arguments{
\item{src}{png file name}

\item{keep_image}{if TRUE, keep image arrays, otherwise discarded}

\item{...}{other arguments for \code{png::readPNG}}
}
\value{
\code{kgschart} class object
}
\description{
KGS Rank Graph Parser
}
\examples{
x <- kgschart(system.file("extdata/leela-ja_JP.png",
                          package = "kgschart"))
}

