% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{kgschart}
\alias{kgschart}
\title{KGS Rank Graph Parser}
\usage{
kgschart(src, ...)
}
\description{
KGS Rank Graph Parser
}
\examples{
x <- kgschart(system.file("extdata/leela-ja_JP.png",
                          package = "kgschart"))
}

