% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{kgschart}
\alias{kgschart}
\title{KGS Rank Graph Parser}
\usage{
kgschart(src, keep_image = FALSE, ...)
}
\arguments{
\item{src}{png file name}

\item{keep_image}{if \code{TRUE}, keep image arrays, otherwise discarded}

\item{...}{other arguments for \code{\link[png]{readPNG}}}
}
\value{
\code{kgschart} class object
}
\description{
Parse a KGS rank graph and recover undelining numeric data.
}
\examples{
x <- kgschart(system.file("extdata/leela-ja_JP.png",
                          package = "kgschart"))
head(x$data)
\dontrun{
plot(x)}
}

