% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{plot.kgschart}
\alias{plot.kgschart}
\title{Plot Method for kgschart object}
\usage{
\method{plot}{kgschart}(x, y = NULL, ...)
}
\arguments{
\item{x}{object}

\item{y}{not in use}

\item{...}{not in use}
}
\value{
\code{gtable} object
}
\description{
Plot Method for kgschart object
}

