% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgschart.R
\name{plot.kgschart}
\alias{plot.kgschart}
\title{Plot method for kgschart object}
\usage{
\method{plot}{kgschart}(x, y = NULL, image = FALSE, separate = FALSE, ...)
}
\arguments{
\item{x}{\code{kgschart} class object}

\item{y}{not in use}

\item{image}{if \code{TRUE}, draw the image, otherwise plot the data}

\item{separate}{if \code{TRUE}, and \code{image} is TRUE, draw the image separated by
component. Useful for debugging.}

\item{...}{not in use}
}
\value{
\code{gtable} object if \code{image} and \code{separate} are both \code{TRUE},
\code{ggplot} object otherwise
}
\description{
Plot data or image
}

